move safe stuff to own file
