+++ package.json
@@ -1,9 +1,9 @@
 {
-  "name": "nearly-safe",
+  "name": "near-safe",
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "ts-node index.ts",
+    "start": "ts-node src/index.ts",
     "fmt": "prettier -w ."
   },
   "dependencies": {
